==> notebooks/model-building/01_data_prep.ipynb <==
import numpy as np
from matplotlib import pyplot as plt
import pandas as pd
import seaborn as sns

#   load data
df = pd.read_csv("..\\data\\02_cleanData_properties_data.csv", index_col="ID")
df.shape

df.drop_duplicates(inplace=True)
df.shape

columns_to_take = [
    "type", "region", "province", "district", "postalCode", "localityType", 
    "bedroomCount", "netHabitableSurface", "condition", 
    "bathroomCount", "showerRoomCount", "toiletCount", 
    "hasLift", "hasGarden", "hasTerrace", 
    "gardenSurface", "terraceSurface",
    "fireplaceExists", "hasSwimmingPool", "hasAirConditioning", 
    "price" 
    ]
    #"kitchen", "pricePerMeter", "constructionYear", 
    #"parkingCountIndoor", "parkingCountOutdoor", "parkingCountClosedBox", 
    #"primaryEnergyConsumptionPerSqm", "epcScore", "hasDoubleGlazing", 
df = df[columns_to_take]
df.info()

df.isna().sum()

df["typeNum"] = df["type"].apply(lambda x: 1 if x == "HOUSE" else 2)
df[["type", "typeNum"]].head()

df.groupby(["typeNum"])["typeNum"].count()

df["regionNum"] = None
df.loc[df["region"] == "Brussels", "regionNum"] = 1
df.loc[df["region"] == "Flanders", "regionNum"] = 2
df.loc[df["region"] == "Wallonie", "regionNum"] = 3

df["provinceNum"] = None
df.loc[df["province"] == "Brussels", "provinceNum"] = 1
df.loc[df["province"] == "Antwerp", "provinceNum"] = 2
df.loc[df["province"] == "East Flanders", "provinceNum"] = 3
df.loc[df["province"] == "Flemish Brabant", "provinceNum"] = 4
df.loc[df["province"] == "Limburg", "provinceNum"] = 5
df.loc[df["province"] == "West Flanders", "provinceNum"] = 6
df.loc[df["province"] == "Hainaut", "provinceNum"] = 7
df.loc[df["province"] == "Liège", "provinceNum"] = 8
df.loc[df["province"] == "Luxembourg", "provinceNum"] = 9
df.loc[df["province"] == "Namur", "provinceNum"] = 10
df.loc[df["province"] == "Walloon Brabant", "provinceNum"] = 11

df.groupby(["regionNum", "provinceNum"])["typeNum"].count()

df.drop(df[df["bedroomCount"] > 12].index, inplace=True)
df.groupby("bedroomCount")["typeNum"].sum()

df.dropna(subset=['netHabitableSurface'], inplace=True)
df["netHabitableSurface"].isna().sum()

df["condition"].fillna(value="UNKNOWN", inplace=True)
df["conditionNum"] = None
df.loc[df["condition"] == "AS_NEW", "conditionNum"] = 1
df.loc[df["condition"] == "JUST_RENOVATED", "conditionNum"] = 2
df.loc[df["condition"] == "GOOD", "conditionNum"] = 3
df.loc[df["condition"] == "TO_BE_DONE_UP", "conditionNum"] = 4
df.loc[df["condition"] == "TO_RENOVATE", "conditionNum"] = 5
df.loc[df["condition"] == "TO_RESTORE", "conditionNum"] = 6
df.loc[df["condition"] == "UNKNOWN", "conditionNum"] = 0
df.groupby("conditionNum")["typeNum"].sum()

df["bathroomCount"].fillna(value=0, inplace=True)
df.loc[df["showerRoomCount"] < 1, "showerRoomCount"] = 0
df.loc[df["showerRoomCount"] > 14, "showerRoomCount"] = 0
df["toiletCount"].fillna(value=0, inplace=True)

# df.groupby("bathroomCount")["type"].count()
# df.groupby("showerRoomCount")["type"].count()
# df.groupby("toiletCount")["type"].count()

binary_to_fix = ["hasLift", "hasGarden", "hasTerrace", "fireplaceExists", "hasSwimmingPool", "hasAirConditioning"]
for btf in binary_to_fix:
    df[btf].fillna(value=False, inplace=True)
    df[btf] = df[btf].apply(lambda x: 1 if x == True else 0)


df[["gardenSurface", "terraceSurface"]].isna().sum()

df["gardenSurface"].fillna(value=0, inplace=True)
df["terraceSurface"].fillna(value=0, inplace=True)

df.groupby("hasAirConditioning")["type"].count()

new_order = [
    "type", "typeNum", "region", "regionNum", "province", "provinceNum", "district", "postalCode", 
    "localityType", "bedroomCount", "netHabitableSurface", "condition", "conditionNum",
    "bathroomCount", "showerRoomCount", "toiletCount", 
    "hasLift", "fireplaceExists", "hasSwimmingPool", "hasAirConditioning", 
    "hasGarden", "gardenSurface", "hasTerrace", "terraceSurface",
    "price"
    ]
df = df[new_order]

df.shape

df.isna().sum()

#   save new csv
path_to_save = "..\\data\\03_ml_prep_data_drop_duplicate.csv"
df.to_csv(path_to_save, index=True)
