==> notebooks/model-building/11_data_prep.ipynb <==
import numpy as np
from matplotlib import pyplot as plt
import pandas as pd
import seaborn as sns

### Load data

#   load data
df = pd.read_csv("..\\data\\02_cleanData_properties_data.csv", index_col="ID")
df.shape

### Drop duplicates

df.drop_duplicates(inplace=True)
df.shape

### Drop missing values

df.dropna(subset=['netHabitableSurface'], inplace=True)
df["netHabitableSurface"].isna().sum()

df.dropna(subset=['price'], inplace=True)
df["price"].isna().sum()

df.dropna(subset=['bedroomCount'], inplace=True)
df["bedroomCount"].isna().sum()

df.shape

### Subset data

"""columns_to_take = [
    "type", "region", "province", "district", "postalCode", "localityType",
    "bedroomCount", "netHabitableSurface", "condition", "epcScore",
    "bathroomCount", "showerRoomCount", "toiletCount", 
    "hasLift", "fireplaceExists", "hasSwimmingPool", "hasAirConditioning", 
    "hasGarden", "hasTerrace", "gardenSurface", "terraceSurface", "land",
    "price" 
    ] 
df = df[columns_to_take]
df.info()"""

### Check null values

df.isna().sum()

df.groupby("localityType")["type"].count()

### Remove/map outliers

df.drop(df[df["bedroomCount"] > 12].index, inplace=True)
#df.groupby("bedroomCount")["type"].count()

df.loc[df["showerRoomCount"] < 1, "showerRoomCount"] = 0
df.loc[df["showerRoomCount"] > 14, "showerRoomCount"] = 14

df.loc[df["gardenSurface"] < 4, "gardenSurface"] = 0
df.loc[df["gardenSurface"] > 1000, "gardenSurface"] = 1000

df.loc[df["terraceSurface"] > 100, "terraceSurface"] = 100

df.loc[df["land"] < 10, "land"] = 0
df.loc[df["land"] > 50000, "land"] = 50000

### Fill null values

df["condition"].fillna(value="UNKNOWN", inplace=True)
#df.groupby("condition")["type"].count()

df["bathroomCount"].fillna(value=0, inplace=True)
df["toiletCount"].fillna(value=0, inplace=True)

# df.groupby("bathroomCount")["type"].count()
# df.groupby("showerRoomCount")["type"].count()
# df.groupby("toiletCount")["type"].count()

binary_to_fix = ["hasLift", "hasGarden", "hasTerrace", "fireplaceExists", "hasSwimmingPool", "hasAirConditioning"]
for btf in binary_to_fix:
    df[btf].fillna(value=False, inplace=True)
    df[btf] = df[btf].apply(lambda x: 1 if x == True else 0)

# df.groupby("hasAirConditioning")["type"].count()

#df[["gardenSurface", "terraceSurface"]].isna().sum()

df["gardenSurface"].fillna(value=0, inplace=True)
#df.groupby("gardenSurface")["type"].count()

df["terraceSurface"].fillna(value=0, inplace=True)
#df.groupby("terraceSurface")["type"].count()

df["epcScore"].fillna(value="UNKNOWN", inplace=True)
#df.groupby("epcScore")["type"].count()

df["land"].fillna(value=0, inplace=True)
#df.groupby("land")["type"].count()

new_order = [
    "type", "region", "province", "district", "postalCode", "localityType",
    "bedroomCount", "netHabitableSurface", "condition", "epcScore",
    "bathroomCount", "showerRoomCount", "toiletCount", 
    "hasLift", "fireplaceExists", "hasSwimmingPool", "hasAirConditioning", 
    "hasGarden", "hasTerrace", "gardenSurface", "terraceSurface", "land",
    "price"
    ]
df = df[new_order]

df.isna().sum()

#   save new csv
path_to_save = "..\\data\\04_ml_prep_data_drop_duplicate.csv"
df.to_csv(path_to_save, index=True)
