==> notebooks/model-building/21_train_models.ipynb <==
import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.tree import DecisionTreeRegressor

import xgboost as xgb
from sklearn.metrics import mean_squared_error as MSE

### Load data

#   load data
df = pd.read_csv("..\\data\\05_ml_data.csv", index_col="ID")

df.shape

X = df.drop("price", axis=1).to_numpy()
y = df["price"].to_numpy()
print(X.shape, y.shape)

### Split data

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0)

print("Shape of X_train: ", X_train.shape)
print("Shape of X_test: ", X_test.shape)
print("Shape of y_train: ", y_train.shape)
print("Shape of y_test: ", y_test.shape)

### Linear regression

model = LinearRegression()
model.fit(X_train, y_train)

model.score(X_train, y_train)

model.score(X_test, y_test)

### Decision tree

dtr = DecisionTreeRegressor(max_depth=12, min_samples_leaf=20)
dtr.fit(X_train, y_train)

dtr.score(X_train, y_train)

dtr.score(X_test, y_test)

### XGB Regressor

xgreg = xgb.XGBRegressor()
xgreg.fit(X_train, y_train)

xgreg.score(X_train, y_train)

xgreg.score(X_test, y_test)
